+++ src/with-options/create-options.js
@@ -1,0 +1,5 @@
+export const createOptions = () => ({
+  scope: '',
+  ensureTargets: [],
+  ensureClasses: [],
+});
