+++ src/with-options/index.js
@@ -1,0 +1,2 @@
+export * from './options-helpers';
+export * from './with-options';
